backend/requirements.txt
flask
flask-cors
groq
Pillow
python-dotenv
